flask/requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.2
python-dateutil==2.8.2 